+++ src/layouts/InterfaceLayout/containers/SwapContainer/SwapContainer.vue
@@ -18,31 +18,24 @@
 
     <div class="title-block">
       <interface-container-title :title="$t('common.swap')" />
-      <div class="buy-eth">
-        <a href="https://ccswap.myetherwallet.com" target="_blank">
-          <span>{{ $t('interface.buyEth') }}</span>
-          <img :src="images.visaMaster" />
-        </a>
-      </div>
     </div>
 
     <div class="send-form">
       <div class="form-block amount-to-address">
         <div class="amount">
-          <div class="title-container">
-            <div class="title title-and-copy">
-              <h4>{{ $t('common.from') }}</h4>
-              <p
-                v-if="tokenBalances[fromCurrency] > 0"
-                class="all-button prevent-user-select"
-                @click="swapAll"
-              >
-                {{ $t('common.totalBalance') }}
-              </p>
-            </div>
+          <div class="title title-and-copy">
+            <h4>{{ $t('common.from') }}</h4>
+            <p
+              v-if="tokenBalances[fromCurrency] > 0"
+              class="all-button prevent-user-select"
+              @click="swapAll"
+            >
+              {{ $t('common.totalBalance') }}
+            </p>
           </div>
           <swap-currency-picker
             :currencies="fromArray"
+            :override-currency="overrideFrom"
             :from-source="true"
             page="SwapContainerFrom"
             @selectedCurrency="setFromCurrency"
@@ -51,8 +44,8 @@
             <input
               v-model="fromValue"
               type="number"
-              name=""
-              value=""
+              name
+              value
               placeholder="Deposit Amount"
               @input="amountChanged('from')"
             />
@@ -62,18 +55,19 @@
             <p v-if="notEnough && !fromBelowMinAllowed">
               {{ $t('common.dontHaveEnough') }}
             </p>
-            <p v-if="!fromBelowMinAllowed">&nbsp;</p>
             <p v-if="fromAboveMaxAllowed">{{ fromAboveMaxAllowed }}</p>
-            <p v-else>&nbsp;</p>
           </div>
         </div>
-        <div class="exchange-icon"><img :src="images.swap" /></div>
+        <div class="exchange-icon" @click="flipCurrencies">
+          <img :src="images.swap" />
+        </div>
         <div class="amount">
           <div class="title">
             <h4>{{ $t('common.to') }}</h4>
           </div>
           <swap-currency-picker
             :currencies="toArray"
+            :override-currency="overrideTo"
             :from-source="false"
             page="SwapContainerTo"
             @selectedCurrency="setToCurrency"
@@ -82,20 +76,19 @@
             <input
               v-model="toValue"
               type="number"
-              name=""
-              value=""
+              name
+              value
               placeholder="Received Amount"
               @input="amountChanged('to')"
             />
           </div>
           <div class="error-message-container">
             <p v-if="toBelowMinAllowed">{{ toBelowMinAllowed }}</p>
-            <p v-else>&nbsp;</p>
             <p v-if="toAboveMaxAllowed">{{ toAboveMaxAllowed }}</p>
-            <p v-else>&nbsp;</p>
           </div>
         </div>
       </div>
+      <!-- form-block amount-to-address -->
     </div>
 
     <div class="send-form">
@@ -149,6 +142,8 @@
         :no-providers-pair="noProvidersPair"
         :loading-data="loadingData"
         :providers-found="providersFound"
+        :provider-selected="selectedProvider"
+        :switch-currency-order="switchCurrencyOrder"
         @selectedProvider="setSelectedProvider"
       />
     </div>
@@ -235,7 +230,8 @@ export default {
       validAddress: true,
       swap: new Swap(providers, {
         network: this.$store.state.network.type.name,
-        web3: this.$store.state.web3
+        web3: this.$store.state.web3,
+        getRateForUnit: true
       }),
       images: {
         kybernetowrk: ImageKybernetowrk,
@@ -256,7 +252,10 @@ export default {
       providersFound: [],
       tempStatuses: [],
       haveProviderRates: false,
-      loadingError: false
+      loadingError: false,
+      overrideFrom: {},
+      overrideTo: {},
+      switchCurrencyOrder: false
     };
   },
   computed: {
@@ -376,7 +375,7 @@ export default {
           new BigNumber(enteredVal)
         );
       } else if (this.fromCurrency === this.baseCurrency) {
-        return +this.fromValue >= this.account.balance;
+        return new BigNumber(this.account.balance).lt(this.fromValue);
       }
       return false;
     }
@@ -416,6 +415,21 @@ export default {
     this.currentAddress = this.wallet.getChecksumAddressString();
   },
   methods: {
+    flipCurrencies() {
+      this.switchCurrencyOrder = true;
+      const origTo = this.toValue;
+      this.fromCurrency = this.currencyDetails.to.symbol;
+      this.toCurrency = this.currencyDetails.from.symbol;
+      this.overrideFrom = this.currencyDetails.to;
+      this.overrideTo = this.currencyDetails.from;
+      this.updateRateEstimate(
+        this.fromCurrency,
+        this.toCurrency,
+        origTo,
+        'from'
+      );
+      this.switchCurrencyOrder = false;
+    },
     setSelectedProvider(provider) {
       this.selectedProvider = this.providerList.find(entry => {
         return entry.provider === provider;
@@ -580,11 +594,17 @@ export default {
                 return {
                   provider: entry.provider,
                   fromCurrency,
-                  fromValue: this.fromValue, // todo uncomment after dev
+                  fromValue: this.fromValue,
                   toCurrency,
                   rate: +entry.rate,
                   minValue: entry.minValue || 0,
-                  maxValue: entry.maxValue || 0
+                  maxValue: entry.maxValue || 0,
+                  computeConversion: function(_fromValue) {
+                    return new BigNumber(_fromValue)
+                      .times(this.rate)
+                      .toFixed(6)
+                      .toString(10);
+                  }
                 };
               }
             }),
@@ -605,7 +625,7 @@ export default {
             providerDetails: providerDetails,
             fromValue: this.fromValue,
             toValue: this.toValue,
-            toAddress: this.toAddress,
+            toAddress: this.toAddress || this.currentAddress,
             fromAddress: this.currentAddress,
             refundAddress: this.swap.isToken(providerDetails.fromCurrency)
               ? this.currentAddress
@@ -634,17 +654,17 @@ export default {
         this.$refs.swapConfirmation.$refs.swapconfirmation.hide();
         this.$refs.swapSendTo.$refs.swapconfirmation.hide();
         this.finalizingSwap = false;
-        // eslint-disable-next-line
+        // eslint-disable-next-line no-console
         console.error(e);
         errorLogger(e);
       }
     },
     resetSwapState() {
-      this.toAddress = '';
+      // this.toAddress = '';
       this.fromCurrency = this.baseCurrency;
-      this.toCurrency = 'BTC';
+      // this.toCurrency = 'BTC';
       this.fromValue = 1;
-      this.toValue = 1;
+      this.toValue = 0;
     }
   }
 };