+++ src/layouts/InterfaceLayout/containers/SwapContainer/components/ProvidersRadioSelector/ProvidersRadioSelector.vue
@@ -1,10 +1,13 @@
 <template>
   <div class="providers-radio-selector">
+    <!-- =========================================================================== -->
     <div v-show="providerData.length > 0" class="radio-button-container">
       <ul>
         <li
           v-for="(provider, idx) in providerData"
           :key="provider.provider + idx"
+          :class="provider.provider"
+          class="providers"
         >
           <div class="mew-custom-form__radio-button">
             <input
@@ -15,6 +18,7 @@
               name="provider"
               @input="setSelectedProvider(provider.provider)"
             />
+
             <label :for="provider.provider" />
           </div>
           <div class="provider-image">
@@ -25,11 +29,34 @@
               minCheck(provider) || maxCheck(provider) ? 'invalid-min-max' : ''
             ]"
           >
-            {{ normalizedRateDisplay(provider) }}
+            {{
+              formatRateDisplay(
+                fromValue,
+                provider.fromCurrency,
+                provider.computeConversion(fromValue),
+                provider.toCurrency
+              )
+            }}
+            <div class="show-mobile">
+              <p
+                v-for="note in minNote(provider)"
+                :key="note.key"
+                :class="[minCheck(provider) ? 'error-message-container' : '']"
+              >
+                {{ note }}
+              </p>
+              <p :class="[maxCheck(provider) ? 'error-message-container' : '']">
+                {{ maxNote(provider) }}
+              </p>
+            </div>
           </div>
-          <div>
-            <p :class="[minCheck(provider) ? 'error-message-container' : '']">
-              {{ minNote(provider) }}
+          <div class="show-desktop">
+            <p
+              v-for="note in minNote(provider)"
+              :key="note.key"
+              :class="[minCheck(provider) ? 'error-message-container' : '']"
+            >
+              {{ note }}
             </p>
             <p :class="[maxCheck(provider) ? 'error-message-container' : '']">
               {{ maxNote(provider) }}
@@ -38,6 +65,24 @@
         </li>
       </ul>
     </div>
+    <!-- Animation while retrieving rates for available providers when switching to and from currencies-->
+    <div
+      v-show="switchCurrencyOrder"
+      class="radio-button-container animated-background"
+    >
+      <ul>
+        <li v-for="provider in providersFound" :key="provider">
+          <div class="mew-custom-form__radio-button">
+            <input type="radio" name="provider" /> <label :for="provider" />
+          </div>
+          <div class="provider-image">
+            <img :src="providerLogo(provider)" />
+          </div>
+          <div class="background-masker" />
+        </li>
+      </ul>
+    </div>
+    <!-- Animation while retrieving rates for available providers -->
     <div
       v-show="loadingData"
       class="radio-button-container animated-background"
@@ -55,6 +100,7 @@
       </ul>
     </div>
     <!-- Animation while retrieving the supporting providers rates -->
+    <!-- =========================================================================== -->
     <div
       v-show="loadingProviderRates"
       class="radio-button-container animated-background"
@@ -71,6 +117,7 @@
       </ul>
     </div>
     <!-- Message When Error Seems to have occured while retrieving rate -->
+    <!-- =========================================================================== -->
     <div
       v-show="loadingProviderError && !noAvaliableProviders"
       class="radio-button-container animated-background"
@@ -91,6 +138,7 @@
       </ul>
     </div>
     <!-- Message when no valid provider is found for the selected pair -->
+    <!-- =========================================================================== -->
     <div v-show="noAvaliableProviders" class="radio-button-container">
       <ul>
         <li>
@@ -105,6 +153,7 @@
         </li>
       </ul>
     </div>
+    <!-- =========================================================================== -->
   </div>
 </template>
 
@@ -149,6 +198,12 @@ export default {
         return [];
       }
     },
+    providerSelected: {
+      type: Object,
+      default: function() {
+        return {};
+      }
+    },
     fromValue: {
       type: Number,
       default: 0
@@ -156,6 +211,10 @@ export default {
     toValue: {
       type: Number,
       default: 0
+    },
+    switchCurrencyOrder: {
+      type: Boolean,
+      default: false
     }
   },
   data() {
@@ -185,6 +244,12 @@ export default {
       return +this.fromValue > details.maxValue && details.maxValue > 0;
     },
     setSelectedProvider(provider) {
+      const providerEls = document.getElementsByClassName('providers');
+      Array.prototype.forEach.call(providerEls, function(el) {
+        el.classList.remove('radio-selected');
+      });
+      const clickedEl = document.getElementsByClassName(provider)[0];
+      clickedEl.classList.add('radio-selected');
       this.$emit('selectedProvider', provider);
     },
     providerLogo(name) {
@@ -193,25 +258,29 @@ export default {
     minNote(details) {
       if (details.minValue > 0) {
         if (details.provider === providerNames.bity) {
-          return `From Min.: ${details.minValue} ${
-            details.fromCurrency
-          } & To Min.: ${details.minValue} ${details.toCurrency}`;
+          return [
+            `${details.minValue} ${details.fromCurrency} (From Min.)`,
+            `${details.minValue} ${details.toCurrency} (From Min.)`
+          ];
         }
-        return `Minimum: ${details.minValue} ${details.fromCurrency}`;
+        return [`${details.minValue} ${details.fromCurrency} (From Min.)`];
       }
       return '';
     },
     maxNote(details) {
       if (details.maxValue > 0) {
         if (details.provider === providerNames.bity) {
-          return `Maximum: ${details.maxValue} ${details.fromCurrency}`;
+          return `${details.maxValue} ${details.fromCurrency} (Max.)`;
         }
-        return `Maximum: ${details.maxValue} ${details.fromCurrency}`;
+        return `${details.maxValue} ${details.fromCurrency} (Max.)`;
       }
       return '';
     },
+    formatRateDisplay(fromValue, fromCurrency, toValue, toCurrency) {
+      return `${fromValue} ${fromCurrency} = ${toValue} ${toCurrency}`;
+    },
     normalizedRateDisplay(source) {
-      const toValue = this.valueForRate(source.fromValue, source.rate);
+      const toValue = this.valueForRate(this.fromValue, source.rate);
       return `${source.fromValue} ${source.fromCurrency} = ${toValue} ${
         source.toCurrency
       }`;
@@ -221,6 +290,14 @@ export default {
         .times(rate)
         .toFixed(6)
         .toString(10);
+    },
+    withDefaultSelectedProvider(provider, idx) {
+      return (
+        provider.provider === this.providerSelected.provider ||
+        (!this.providerSelected.provider && idx === 0)
+      );
+      // if (this.providerSelected === '' && idx === 0) {
+      // }
     }
   }
 };