+++ src/partners/helpers/index.js
@@ -1,11 +1,13 @@
 import {
   bestRateForQuantity,
   bestProviderForQuantity,
-  isValidEntry,
-  checkInvalidOrMissingValue,
   dynamicSortMultiple
 } from './sortAndIdentify';
 import * as utils from './utils';
+import qrcodeBuilder from './qrCodeBuilder';
+
+const isValidEntry = utils.isValidEntry;
+const checkInvalidOrMissingValue = utils.checkInvalidOrMissingValue;
 
 export {
   utils,
@@ -13,5 +15,6 @@ export {
   bestProviderForQuantity,
   isValidEntry,
   checkInvalidOrMissingValue,
-  dynamicSortMultiple
+  dynamicSortMultiple,
+  qrcodeBuilder
 };