+++ src/wallets/web3-provider/methods/eth_sendTransaction.js
@@ -11,13 +11,13 @@ import BigNumber from 'bignumber.js';
 const setEvents = (promiObj, tx, dispatch) => {
   promiObj
     .once('transactionHash', hash => {
-      dispatch('addNotification', [tx.from, hash, 'Transaction Hash']);
+      dispatch('addNotification', ['Hash', tx.from, tx, hash]);
     })
-    .on('receipt', res => {
-      dispatch('addNotification', [tx.from, res, 'Transaction Receipt']);
+    .once('receipt', res => {
+      dispatch('addNotification', ['Receipt', tx.from, tx, res]);
     })
     .on('error', err => {
-      dispatch('addNotification', [tx.from, err, 'Transaction Error']);
+      dispatch('addNotification', ['Error', tx.from, tx, err]);
     });
 };
 export default async (
@@ -44,31 +44,7 @@ export default async (
   getSanitizedTx(tx)
     .then(_tx => {
       if (store.state.wallet.identifier === WEB3_WALLET) {
-        eventHub.$emit(EventNames.SHOW_WEB3_CONFIRM_MODAL, _tx, _response => {
-          const _promiObj = store.state.web3.eth.sendSignedTransaction(
-            _response.rawTransaction
-          );
-          _promiObj
-            .once('transactionHash', hash => {
-              const localStoredObj = locStore.get(
-                utils.sha3(store.state.wallet.getChecksumAddressString())
-              );
-              locStore.set(
-                utils.sha3(store.state.wallet.getChecksumAddressString()),
-                {
-                  nonce: utils.toHex(
-                    new BigNumber(localStoredObj.nonce).plus(1).toFixed()
-                  ),
-                  timestamp: localStoredObj.timestamp
-                }
-              );
-              res(null, toPayload(payload.id, hash));
-            })
-            .on('error', err => {
-              res(err);
-            });
-          setEvents(_promiObj, _tx, store.dispatch);
-        });
+        eventHub.$emit(EventNames.SHOW_WEB3_CONFIRM_MODAL, _tx);
       } else {
         eventHub.$emit(EventNames.SHOW_TX_CONFIRM_MODAL, _tx, _response => {
           const _promiObj = store.state.web3.eth.sendSignedTransaction(